Update video quality settings
Selecting fixed resolution makes sense only for HLS streams.
Supersedes #290.
