+++ src/resources/lib/utils.py
@@ -1,6 +1,9 @@
+import re
 from typing import Any
 from typing import Callable
+from typing import List
 from typing import Type
+from typing import Union
 
 import xbmcaddon
 import xbmcgui
@@ -31,6 +34,16 @@ class cached_property:  # noqa
             return self
         value = obj.__dict__[self.func.__name__] = self.func(obj)
         return value
+
+
+def natural_sort(lines: List[str]) -> List[str]:
+    def convert(text: str) -> Union[str, int]:
+        return int(text) if text.isdigit() else text.lower()
+
+    def alphanum_key(key: str) -> List[Union[str, int]]:
+        return [convert(c) for c in re.split("([0-9]+)", key)]
+
+    return sorted(lines, key=alphanum_key)
 
 
 localize = xbmcaddon.Addon().getLocalizedString
